+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### [2025-09-13]
 - Added capabilities to download `Ubuntu 22.04` cloud image and build kvm guest machines.
+- Added capabilities to create KV networks. Review "Create Additional KVM Networks" section for more information.